changeing to POST

+++ administrator/components/com_media/src/Controller/AdaptiveImageController.php
@@ -66,11 +66,18 @@ class AdaptiveImageController extends BaseController
 				$this->cropImage($imgPath, $widths);
 				return true;
 			case "deleteFocus" :
-				$imgPath = $this->input->getString('path');
-				$storage = new JSONFocusStore;
-				$storage->deleteFocus($imgPath);
-				$storage->deleteResizedImages($imgPath);
-				return true;
+				if ($this->input->getMethod() != 'POST')
+				{
+					return false;
+				}
+				else
+				{
+					$imgPath = $this->input->getString('path');
+					$storage = new JSONFocusStore;
+					$storage->deleteFocus($imgPath);
+					$storage->deleteResizedImages($imgPath);
+					return true;
+				}
 			default :
 				return false;
 		}
